backend/requirements.txt
# Flask e extensões essenciais
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0
Flask-Mail==0.9.1
Flask-Limiter==3.5.0

# Banco de dados
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23

# Segurança e criptografia  
bcrypt==4.1.1
flask-bcrypt==1.0.1
cryptography==41.0.7
python-dotenv==1.0.0

# Validação e serialização
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0
Werkzeug==3.0.1

# Utilitários
requests==2.31.0
python-dateutil==2.8.2
pytz==2023.3
celery==5.3.4
redis==5.0.1

# Upload e storage
Pillow==10.1.0
boto3==1.34.0

# Desenvolvimento e testes
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0
factory-boy==3.3.0
faker==20.1.0

# Produção
gunicorn==21.2.0
python-json-logger==2.0.7

# Análise e documentação
flask-smorest==0.42.3
apispec==6.3.0